+++ CHANGES.md
@@ -1,6 +1,12 @@
 1.0.5
 =====
 
+* [#60](https://github.com/serokell/universum/issues/60):
+  Reexport `Natural` type from `Numeric.Natural` module.
+* [#118](https://github.com/serokell/universum/issues/118):
+  Reexport `Type` from `Data.Kind` module.
+* [#130](https://github.com/serokell/universum/issues/130):
+  Merge `ToList` and `Container` type classes into single type class `Container`.
 * [#15](https://github.com/serokell/universum/issues/15):
   Add `?:` function to `Universum.Monad.Maybe`.
 * [#128](https://github.com/serokell/universum/issues/128):